requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
astroid==3.0.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
bcrypt==3.2.0
blinker==1.6.3
Brlapi==0.8.3
cachelib==0.10.2
caffeine==2.9.8
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.6.1
dbus-python==1.2.18
defer==1.0.6
dill==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
duplicity==0.8.21
ewmh==0.1.6
exceptiongroup==1.1.3
fasteners==0.14.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-OAuthlib==0.9.6
frozenlist==1.4.0
future==0.18.2
greenlet==3.0.1
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isort==5.12.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
keyring==23.5.0
langchain==0.0.324
langsmith==0.0.52
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.26.1
oauthlib==2.1.0
olefile==0.46
openai==0.28.1
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.11.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.11.1
pycups==2.0.1
pydantic==2.4.2
pydantic_core==2.10.1
PyGObject==3.42.1
PyJWT==2.3.0
pylint==3.0.2
pymacaroons==0.13.0
pymongo==4.5.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.1.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
systemd-python==234
tenacity==8.2.3
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
